chore: rename folder for improved naming clarity
