--- src/components/context/index.ts
@@ -1,4 +1,0 @@
-import ThemeProvider from './ThemeProvider';
-import ThemeToggle from './ThemeToggle';
-
-export { ThemeProvider, ThemeToggle };
